separate files for server and serverTest
