+++ server.py
@@ -4,7 +4,7 @@ from flask import Flask
 from flask_socketio import SocketIO, emit
 
 # create Flask object
-app = Flask(__name__,template_folder='./static',static_folder='./static')
+app = Flask(__name__)
 app.config['SECRET_KEY'] = 'secret!'
 socketio = SocketIO(app)
 
@@ -21,27 +21,5 @@ def handle_stroke(json):
     print("received json: " + str(json))
     emit('receive-update', "myBoard")
 
-
-class TestSocketIO(unittest.TestCase):
-    def test_connect(self):
-        client1 = socketio.test_client(app)
-        client2 = socketio.test_client(app)
-        self.assertTrue(client1.is_connected())
-        self.assertTrue(client2.is_connected())
-        self.assertNotEqual(client1.sid, client2.sid)
-        client1.disconnect()
-        self.assertFalse(client1.is_connected())
-        self.assertTrue(client2.is_connected())
-        client2.disconnect()
-        self.assertFalse(client2.is_connected())
-
-    def test_stroke(self):
-        client = socketio.test_client(app)
-        client.emit('send-stroke', 'myStroke')
-        received_data = client.get_received()
-        myData = received_data[0]['args'][0]
-        self.assertEqual(myData, "myBoard")
-
-
-if __name__ == '__main__':
-    unittest.main()
+if __name__ == "__main__":
+    socketio.run(app, port=8080, debug=True)
